requirements.txt
Django>=3.1.4,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
Pillow>=8.0.1,<8.1.0
# flake8>=3.8.4,<3.9.0
channels>=3.0.3,<3.1.0
channels-redis>=3.2.0,<3.3.0
celery>= 5.0.5,<5.1.0
# cryptography==2.8
cryptography==3.4.6
gunicorn>= 20.0.4,<20.1.0
redis>= 3.5.3,<3.6.0
django-celery-beat>=2.2.0,<2.3.0
django-celery-results>=2.0.1 ,<2.1.0
django-push-notifications>=2.0.0,<2.1.0
daphne==3.0.1
idna<3,>=2.5
# uWSGI>=2.0.19.1,<2.1.0.0

# gunicorn==20.0.4
# uvicorn==0.12.0
# httptools==0.1.1
# uvloop==0.14.0